amend iter impl
