+++ go/cmd/dolt/commands/engine/sql_print.go
@@ -16,16 +16,13 @@ package engine
 
 import (
 	"context"
-	"encoding/hex"
 	"errors"
 	"fmt"
 	"io"
-	"strings"
 	"time"
 
 	"github.com/dolthub/go-mysql-server/sql"
 	"github.com/dolthub/go-mysql-server/sql/types"
-	"github.com/dolthub/vitess/go/sqltypes"
 	"github.com/fatih/color"
 
 	"github.com/dolthub/dolt/go/cmd/dolt/cli"
@@ -129,9 +126,8 @@ func prettyPrintResultsWithSummary(ctx *sql.Context, resultFormat PrintResultFor
 			}
 		}
 
-		// Wrap iterator with binary-to-hex transformation if needed
 		if binaryAsHex {
-			rowIter = newBinaryHexIterator(rowIter, sqlSch)
+			rowIter = sqlutil.NewBinaryHexIterator(rowIter, sqlSch)
 		}
 
 		numRows, err = writeResultSet(ctx, rowIter, wr)
@@ -203,124 +199,6 @@ func printResultSetSummary(numRows int, numWarnings uint16, warningsList string,
 	}
 
 	return nil
-}
-
-// binaryHexIterator wraps a row iterator and transforms binary data to hex format
-type binaryHexIterator struct {
-	inner  sql.RowIter
-	schema sql.Schema
-}
-
-var _ sql.RowIter = (*binaryHexIterator)(nil)
-
-// BinaryAsHexString serves as an indicator that a binary value has been processed by the --binary-as-hex flag iterator
-// into a hex string value.
-type BinaryAsHexString string
-
-// newBinaryHexIterator creates a new iterator that transforms binary data to hex format
-func newBinaryHexIterator(inner sql.RowIter, schema sql.Schema) sql.RowIter {
-	return &binaryHexIterator{
-		inner:  inner,
-		schema: schema,
-	}
-}
-
-// Next returns the next row with binary data transformed to hex format.
-func (iter *binaryHexIterator) Next(ctx *sql.Context) (rowData sql.Row, err error) {
-	rowData, err = iter.inner.Next(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	for i, val := range rowData {
-		if bytesWrapper, ok := val.(sql.BytesWrapper); ok {
-			val, err = bytesWrapper.Unwrap(ctx)
-			if err != nil {
-				return nil, err
-			}
-		}
-
-		hexBytes, err := binaryToUpperHexBytes(val)
-		if err != nil {
-			return nil, err
-		}
-
-		var strBuilder strings.Builder
-		switch iter.schema[i].Type.Type() {
-		case sqltypes.Binary, sqltypes.VarBinary, sqltypes.Blob:
-			strBuilder.Grow(2 + len(hexBytes))
-			strBuilder.WriteByte('0')
-			strBuilder.WriteByte('x')
-			strBuilder.Write(hexBytes)
-			rowData[i] = BinaryAsHexString(strBuilder.String())
-		case sqltypes.Bit:
-			padding := 0
-			if bitType, ok := iter.schema[i].Type.(types.BitType); ok {
-				padding = (int(bitType.NumberOfBits())+3)/4 - len(hexBytes)
-			}
-
-			strBuilder.Grow(2 + padding + len(hexBytes))
-			strBuilder.WriteByte('0')
-			strBuilder.WriteByte('x')
-			for repeat := 0; repeat < padding; repeat++ {
-				strBuilder.WriteByte('0')
-			}
-			strBuilder.Write(hexBytes)
-			rowData[i] = BinaryAsHexString(strBuilder.String())
-		}
-	}
-
-	return rowData, nil
-}
-
-// binaryToUpperHexBytes converts the input |binary| into uppercase hexadecimal bytes. It accepts binary and numeric
-// input types. This is optimized for large byte arrays i.e. sqltypes.Blob, reimplementing a modified version of the hex
-// encoder from Go's standard library to do uppercasing in a single pass.
-func binaryToUpperHexBytes(binary interface{}) ([]byte, error) {
-	upperHexTable := "0123456789ABCDEF"
-	var valBytes []byte
-	switch v := binary.(type) {
-	case []byte:
-		valBytes = v
-	case string:
-		valBytes = []byte(v)
-	case uint64:
-		if v == 0 {
-			return []byte{'0'}, nil
-		}
-		valBytes = make([]byte, 16)
-		// uint64 contains 16 nibbles (4-bit chunks) obtained by & 0xF.
-		// Map each nibble directly to the hex table above.
-		// We assume big-endian and shift right to process from least to greatest bit.
-		index := 15
-		for v > 0 {
-			valBytes[index] = upperHexTable[v&0xF]
-			v >>= 4
-			index--
-		}
-
-		// Index + 1 to avoid leading zeros.
-		return valBytes[index+1:], nil
-	default:
-		return nil, fmt.Errorf("unexpected type %T (%v)", binary, binary)
-	}
-
-	if len(valBytes) == 0 {
-		return []byte{}, nil
-	}
-
-	hexBuffer := make([]byte, hex.EncodedLen(len(valBytes)))
-	for index, valByte := range valBytes {
-		// Each byte contains 2 nibbles (4-bit chunks), map each to upper hex table above.
-		hexBuffer[index*2] = upperHexTable[valByte>>4]
-		hexBuffer[index*2+1] = upperHexTable[valByte&0x0F]
-	}
-	return hexBuffer, nil
-}
-
-// Close closes the wrapped iterator and releases any resources.
-func (iter *binaryHexIterator) Close(ctx *sql.Context) error {
-	return iter.inner.Close(ctx)
 }
 
 // writeResultSet drains the iterator given, printing rows from it to the writer given. Returns the number of rows.
