+++ go/cmd/dolt/commands/sqlserver/queryist_utils.go
@@ -17,13 +17,15 @@ package sqlserver
 import (
 	"context"
 	"crypto/tls"
-	sql2 "database/sql"
+	driversql "database/sql"
 	"fmt"
 	"io"
 	"regexp"
 	"strings"
 
 	"github.com/dolthub/go-mysql-server/sql"
+	gmstypes "github.com/dolthub/go-mysql-server/sql/types"
+	"github.com/dolthub/vitess/go/sqltypes"
 	"github.com/dolthub/vitess/go/vt/sqlparser"
 	"github.com/go-sql-driver/mysql"
 	"github.com/gocraft/dbr/v2"
@@ -32,7 +34,6 @@ import (
 	"github.com/dolthub/dolt/go/cmd/dolt/cli"
 	"github.com/dolthub/dolt/go/libraries/doltcore/doltdb"
 	"github.com/dolthub/dolt/go/libraries/doltcore/servercfg"
-	"github.com/dolthub/dolt/go/libraries/doltcore/sqle/sqlutil"
 	"github.com/dolthub/dolt/go/libraries/utils/argparser"
 	"github.com/dolthub/dolt/go/libraries/utils/filesys"
 )
@@ -93,7 +94,7 @@ func BuildConnectionStringQueryist(ctx context.Context, cwdFS filesys.Filesys, c
 		return nil, err
 	}
 
-	conn := &dbr.Connection{DB: sql2.OpenDB(mysqlConnector), EventReceiver: nil, Dialect: dialect.MySQL}
+	conn := &dbr.Connection{DB: driversql.OpenDB(mysqlConnector), EventReceiver: nil, Dialect: dialect.MySQL}
 
 	gatherWarnings := false
 	queryist := ConnectionQueryist{connection: conn, gatherWarnings: &gatherWarnings}
@@ -131,12 +132,12 @@ func (c ConnectionQueryist) Query(ctx *sql.Context, query string) (sql.Schema, s
 		return nil, nil, nil, err
 	}
 
-	rowIter, err := NewMysqlRowWrapper(rows)
+	rowIter, err := NewConnectionQueryResult(rows)
 	if err != nil {
 		return nil, nil, nil, err
 	}
 
-	if c.gatherWarnings != nil && *c.gatherWarnings == true {
+	if c.gatherWarnings != nil && *c.gatherWarnings {
 		ctx.ClearWarnings()
 
 		re := regexp.MustCompile(`\s+`)
@@ -171,29 +172,38 @@ func (c ConnectionQueryist) QueryWithBindings(ctx *sql.Context, query string, _ 
 	return c.Query(ctx, query)
 }
 
-type MysqlRowWrapper struct {
+type ConnectionQueryResult struct {
 	rows    []sql.Row
 	schema  sql.Schema
 	numRows int
 	curRow  int
 }
 
-var _ sql.RowIter = (*MysqlRowWrapper)(nil)
+var _ sql.RowIter = (*ConnectionQueryResult)(nil)
 
-func NewMysqlRowWrapper(sqlRows *sql2.Rows) (*MysqlRowWrapper, error) {
+func NewConnectionQueryResult(sqlRows *driversql.Rows) (*ConnectionQueryResult, error) {
 	colTypes, err := sqlRows.ColumnTypes()
 	if err != nil {
 		return nil, err
 	}
+
 	schema := make(sql.Schema, len(colTypes))
 	vRow := make([]*string, len(colTypes))
 	iRow := make([]interface{}, len(colTypes))
 	rows := make([]sql.Row, 0)
 	for i, colType := range colTypes {
+		colTypeNullable, ok := colType.Nullable()
+		if !ok {
+			return nil, fmt.Errorf("column '%s' of type '%s' has no nullable", colType.Name(), colType.Name())
+		}
+		sqlType, err := newSqlTypeFromDriverColumnType(colType)
+		if err != nil {
+			return nil, err
+		}
 		schema[i] = &sql.Column{
 			Name:     colType.Name(),
-			Type:     sqlutil.DatabaseTypeNameToSqlType(colType.DatabaseTypeName()),
-			Nullable: true,
+			Type:     sqlType,
+			Nullable: colTypeNullable,
 		}
 		iRow[i] = &vRow[i]
 	}
@@ -218,7 +228,7 @@ func NewMysqlRowWrapper(sqlRows *sql2.Rows) (*MysqlRowWrapper, error) {
 		return nil, err
 	}
 
-	return &MysqlRowWrapper{
+	return &ConnectionQueryResult{
 		rows:    rows,
 		schema:  schema,
 		numRows: len(rows),
@@ -226,11 +236,41 @@ func NewMysqlRowWrapper(sqlRows *sql2.Rows) (*MysqlRowWrapper, error) {
 	}, nil
 }
 
-func (s *MysqlRowWrapper) Schema() sql.Schema {
+// newSqlTypeFromDriverColumnType converts a database/sql ColumnType into a go-mysql-server sql.Type.
+// This mapping is required because the database/sql interface exposes types as strings rather
+// than their underlying wire protocol details. We reconstruct GMS types from the driver metadata
+// to ensure consistent type handling when Dolt acts as a MySQL client.
+func newSqlTypeFromDriverColumnType(columnType *driversql.ColumnType) (sql.Type, error) {
+	typeName := columnType.DatabaseTypeName()
+	typeLength, ok := columnType.Length()
+	if !ok {
+		return nil, fmt.Errorf("column '%s' of type '%s' has no length", columnType.Name(), typeName)
+	}
+	switch typeName {
+	case "binary":
+		return gmstypes.MustCreateBinary(sqltypes.Binary, typeLength), nil
+	case "varbinary":
+		return gmstypes.MustCreateBinary(sqltypes.VarBinary, typeLength), nil
+	case "bit":
+		return gmstypes.MustCreateBitType(uint8(typeLength)), nil
+	case "tinyblob":
+		return gmstypes.TinyBlob, nil
+	case "blob":
+		return gmstypes.Blob, nil
+	case "mediumblob":
+		return gmstypes.MediumBlob, nil
+	case "longblob":
+		return gmstypes.LongBlob, nil
+	default:
+		return gmstypes.LongText, nil
+	}
+}
+
+func (s *ConnectionQueryResult) Schema() sql.Schema {
 	return s.schema
 }
 
-func (s *MysqlRowWrapper) Next(*sql.Context) (sql.Row, error) {
+func (s *ConnectionQueryResult) Next(*sql.Context) (sql.Row, error) {
 	if s.NoMoreRows() {
 		return nil, io.EOF
 	}
@@ -239,11 +279,11 @@ func (s *MysqlRowWrapper) Next(*sql.Context) (sql.Row, error) {
 	return s.rows[s.curRow-1], nil
 }
 
-func (s *MysqlRowWrapper) NoMoreRows() bool {
+func (s *ConnectionQueryResult) NoMoreRows() bool {
 	return s.curRow >= s.numRows
 }
 
-func (s *MysqlRowWrapper) Close(*sql.Context) error {
+func (s *ConnectionQueryResult) Close(*sql.Context) error {
 	s.curRow = s.numRows
 	return nil
 }
