+++ go/libraries/doltcore/sqle/sqlutil/sql_row.go
@@ -16,6 +16,7 @@ package sqlutil
 
 import (
 	"context"
+	"encoding/hex"
 	"errors"
 	"fmt"
 	"strings"
@@ -27,11 +28,14 @@ import (
 	_ "github.com/dolthub/go-mysql-server/sql/variables"
 	"github.com/dolthub/vitess/go/sqltypes"
 
-	"github.com/dolthub/dolt/go/cmd/dolt/commands/engine"
 	"github.com/dolthub/dolt/go/libraries/doltcore/row"
 	"github.com/dolthub/dolt/go/libraries/doltcore/schema"
 	"github.com/dolthub/dolt/go/store/types"
 )
+
+// BinaryAsHexString serves as an indicator that a binary value has been processed by the --binary-as-hex flag iterator
+// into a hex string value.
+type BinaryAsHexString string
 
 // DoltRowToSqlRow constructs a go-mysql-server sql.Row from a Dolt row.Row.
 func DoltRowToSqlRow(doltRow row.Row, sch schema.Schema) (sql.Row, error) {
@@ -228,7 +232,7 @@ func keylessDoltRowFromSqlRow(ctx context.Context, vrw types.ValueReadWriter, sq
 // NULL values are treated as empty strings. Handle nil separately if you require other behavior.
 func SqlColToStr(ctx *sql.Context, sqlType sql.Type, col interface{}) (string, error) {
 	if col != nil {
-		if hexVal, ok := col.(engine.BinaryAsHexString); ok {
+		if hexVal, ok := col.(BinaryAsHexString); ok {
 			return string(hexVal), nil
 		}
 
@@ -258,20 +262,119 @@ func SqlColToStr(ctx *sql.Context, sqlType sql.Type, col interface{}) (string, e
 	return "", nil
 }
 
-// DatabaseTypeNameToSqlType converts a MySQL wire protocol database type name
-// to a go-mysql-server sql.Type. This uses the same type mapping logic as the existing
-// Dolt type system for consistency.
-// TODO: Add support for BLOB types (TINYBLOB, BLOB, MEDIUMBLOB, LONGBLOB) and BIT type
-// as confirmed by testing MySQL 8.4+ binary-as-hex behavior
-func DatabaseTypeNameToSqlType(databaseTypeName string) sql.Type {
-	typeName := strings.ToLower(databaseTypeName)
-	switch typeName {
-	case "binary":
-		return gmstypes.MustCreateBinary(sqltypes.Binary, 255)
-	case "varbinary":
-		return gmstypes.MustCreateBinary(sqltypes.VarBinary, 255)
-	default:
-		// Default to LongText for all other types (as was done before)
-		return gmstypes.LongText
+// binaryHexIterator wraps a row iterator and transforms binary data to hex format
+type binaryHexIterator struct {
+	inner  sql.RowIter
+	schema sql.Schema
+}
+
+var _ sql.RowIter = (*binaryHexIterator)(nil)
+
+// NewBinaryHexIterator creates a new iterator that transforms binary data to hex format
+func NewBinaryHexIterator(inner sql.RowIter, schema sql.Schema) sql.RowIter {
+	return &binaryHexIterator{
+		inner:  inner,
+		schema: schema,
 	}
+}
+
+// Next returns the next row with binary data transformed to hex format.
+func (iter *binaryHexIterator) Next(ctx *sql.Context) (rowData sql.Row, err error) {
+	rowData, err = iter.inner.Next(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	for i, val := range rowData {
+		if bytesWrapper, ok := val.(sql.BytesWrapper); ok {
+			val, err = bytesWrapper.Unwrap(ctx)
+			if err != nil {
+				return nil, err
+			}
+		}
+
+		hexBytes, err := convertBinaryToUpperHexBytes(val)
+		if err != nil {
+			return nil, err
+		}
+
+		var strBuilder strings.Builder
+		switch iter.schema[i].Type.Type() {
+		case sqltypes.Binary, sqltypes.VarBinary, sqltypes.Blob:
+			strBuilder.Grow(2 + len(hexBytes))
+			strBuilder.WriteByte('0')
+			strBuilder.WriteByte('x')
+			strBuilder.Write(hexBytes)
+			rowData[i] = BinaryAsHexString(strBuilder.String())
+		case sqltypes.Bit:
+			padding := 0
+			if bitType, ok := iter.schema[i].Type.(gmstypes.BitType); ok {
+				expectedLen := (int(bitType.NumberOfBits()) + 3) / 4
+				if len(hexBytes) < expectedLen {
+					padding = expectedLen - len(hexBytes)
+				}
+			}
+
+			strBuilder.Grow(2 + padding + len(hexBytes))
+			strBuilder.WriteByte('0')
+			strBuilder.WriteByte('x')
+			for repeat := 0; repeat < padding; repeat++ {
+				strBuilder.WriteByte('0')
+			}
+			strBuilder.Write(hexBytes)
+			rowData[i] = BinaryAsHexString(strBuilder.String())
+		}
+	}
+
+	return rowData, nil
+}
+
+// convertBinaryToUpperHexBytes converts the input |binary| into uppercase hexadecimal bytes. This is optimized for
+// large byte arrays i.e. sqltypes.Blob, reimplementing a modified version of the hex encoder from Go's standard library
+// to do uppercasing in a single pass.
+func convertBinaryToUpperHexBytes(val interface{}) ([]byte, error) {
+	upperHexTable := "0123456789ABCDEF"
+	var valBytes []byte
+	switch v := val.(type) {
+	case []byte:
+		valBytes = v
+	case string:
+		valBytes = []byte(v)
+	case uint64:
+		if v == 0 {
+			return []byte{'0'}, nil
+		}
+		valBytes = make([]byte, 16)
+		// uint64 contains 16 nibbles (4-bit chunks) obtained by & 0xF.
+		// Map each nibble directly to the hex table above.
+		// We assume big-endian and shift right to process from least to greatest bit.
+		index := 15
+		for v > 0 {
+			valBytes[index] = upperHexTable[v&0xF]
+			v >>= 4
+			index--
+		}
+
+		// Index + 1 to avoid leading zeros.
+		return valBytes[index+1:], nil
+	default:
+		return nil, fmt.Errorf("unexpected type %T (%v)", val, val)
+	}
+
+	if len(valBytes) == 0 {
+		return []byte{}, nil
+	}
+
+	hexBuffer := make([]byte, hex.EncodedLen(len(valBytes)))
+	for index, valByte := range valBytes {
+		// Each byte contains 2 nibbles (4-bit chunks), map each to upper hex table above.
+		hexBuffer[index*2] = upperHexTable[valByte>>4]
+		hexBuffer[index*2+1] = upperHexTable[valByte&0x0F]
+	}
+	return hexBuffer, nil
+}
+
+// Close closes the wrapped iterator and releases any resources.
+func (iter *binaryHexIterator) Close(ctx *sql.Context) error {
+	return iter.inner.Close(ctx)
 }
